les13-14: query and mutation, create items & prisma yoga flow
When you want to add a new type, first you add it to datamodel.graphql, then run prisma deploy to push that up yo the prisma service, that brings down a new copy of the prisma.graphql which contains all of our queries and our mutations and all the different filters that we could use.
$ npm run deploy; npm run dev
Then we go into our own schema.graphql which is our public facing api, this is what React app is going to interface with. We can create the ability to create new functions, like create items or query items, then to match up with these mutations and queries and go ahead and write the resolvers to actualy complete all the logic, like sending email, charging credit cards etc.