les29: backend password reset flow
At the prisma.graphql Prisma file take the User type and redefine it in schema.graphql Yoga file.
Promisify takes randomBytes that is a callback function and returns it as a promise function.
At the Mutation.js file we have the requestReset function that takes parent, args, ctx, info and do the followings:
1 - Check if this is the real user
2 - Set a reset token and expiry on that user
3 - Email them that reset token
At the Mutation.js file we have the resetPassword function that takes parent, args, ctx, info and do the followings:
1 - Check if the passwords match
2 - Check if its a legit reset token
3 - Check if its expired
4 - Hash the new password
5 - Save the new password to the user and remove old resetToken fields
6 - Generate JWT
7 - Set the JWT cookie
8 - Return the new user