les31: sending email from accounts, authent & permiss mod
Mailtrap simulates the work of a real SMTP server. It isolates the staging emailing from production and eliminates any possibility of a test email to land in a real customer’s mailbox. The account credentials of mailtrap goes at the variables.env file that ignored grom the git. Then a new file mail.js is taking the credentials using  process.env.MAIL_USER for ex. and exporting them as transport. Also we making a simple mail template html and exporting it as makeANiceEmail.
At the Mutation.js file of resolvers we importing with require the transport and makeANiceEmail. At the requestReset async function and before returning the message we build the email await function for sending that reset token.
So, you can test the password reset with token sending functionality of the application and catch the email sent at the mailtrap mail service for developing purpose.