les32: data relationships from accounts, authent, permiss module
In order to make a relationship connection from Item to User we have to change datamodel.graphql adding at Item type user: User!
But to take effect the changes with npm run deploy no entry have to exist, because of the User!
Then at the Mutation.js at the mutation.createItem await function we create a relationship between the Item and the User.