les40: server side add to cart
At the Graphql datamodel we create a new type, the CartItem that related to both the item and the user. Also we add the   cart: [CartItem!]!   at the User. At Graphql schema we add   cart: [CartItem!]!   also. Because the changes of the Datamodel we need to   npm run deploy. Also, because the changes of the Schema we implement a resolver that lines up with it. At the Mutation.js adding a new async  addToCart(parent, args, ctx, info) { } where:
 1. Make sure a user is signed in
 2. Query the users current cart
 3. Check if that item is already in the user's cart and increment by 1 if it is
 4. If it's not, create a new CartItem for that user