les50: charging cards on the server side
After token sent to the server we are going to charge it. At datamodel.graphql defining a new   type OrderItem { }  and a new   type Order { }   Then going at schema.graphql and adding a mutation createOrder(token: String!): Order!  Then going at the Mutation.js and creating a resolver  createOrder( ){  }  for the mutation we added, and do these:
 1. Query the current user and make sure they are signed in
 2. Recalculate the total for the price
 3. Create the stripe charge
 4. Convert the CartItems to OrderItems
 5. Create the order
 6. Clear the users cart, delete cartItems
 7. Return the order to the client